refactor: [#1551] extract event handler for each udp event
